[JS] Logic separated into multiple files
