+++ dataToJson.js
@@ -1,280 +1,31 @@
 // TODO retirer la fin des pokemons missingno + CAP + pokemon studio
 'use strict';
-const fileSystem = require('fs');
-const { Abilities } = require('./pokemon-showdown/.data-dist/abilities');
-const { AbilitiesText } = require('./pokemon-showdown/.data-dist/text/abilities');
-const { Items } = require('./pokemon-showdown/.data-dist/items');
-const { ItemsText } = require('./pokemon-showdown/.data-dist/text/items');
-const { TypeChart } = require('./pokemon-showdown/.data-dist/typechart');
-const { Moves } = require('./pokemon-showdown/.data-dist/moves');
-const { MovesText } = require('./pokemon-showdown/.data-dist/text/moves');
-const { Pokedex } = require('./pokemon-showdown/.data-dist/pokedex');
-const { PokedexText } = require('./pokemon-showdown/.data-dist/text/pokedex');
-const { Learnsets } = require('./pokemon-showdown/.data-dist/learnsets');
-const { Natures } = require('./pokemon-showdown/.data-dist/natures');
-const { FormatsData } = require('./pokemon-showdown/.data-dist/formats-data');
-const LAST_GEN = 8
 
-const removeParenthesis = string => string.replace(/\(+/g, '').replace(/\)+/g, '');
+const { writeFile } = require('./util');
 
-const log = (name, e) => {
-	if (e) {
-		console.error(`Erreur lors de la création du  fichier ${name}.json :`, e);
-	} else {
-		console.info(`Création du fichier ${name}.json réussi`);
-	}
-};
+const abilities = require('./abilities')
+const items = require('./items')
+const learns = require('./learns')
+const moves = require('./moves')
+const natures = require('./natures')
+const pokemons = require('./pokemon')
+const pokemonTier = require('./pokemonTier')
+const types = require('./types')
 
-// prettier-ignore
-const writeFile = (fileName, values) => fileSystem.writeFile(
-	`json/${fileName}.json`,
-	JSON.stringify(values),
-	e => log(fileName, e)
-);
+writeFile('abilities', abilities)
 
-const pokemonIsStandard = value =>
-	!value.isNonstandard ||
-	value.isNonstandard === 'Past' || // keep pokemons that are not import in gen8
-	value.isNonstandard === 'Gigantamax' || // keep Gmax forms
-	value.isNonstandard === 'Unobtainable' &&  // keep Unobtainable real mons
-	value.tier !== 'Illegal'
+writeFile('pokemons', pokemons)
 
-const abilities = Object.entries(Abilities)
-	.filter(([key, value]) => !value.isNonstandard || value.isNonstandard === 'Past')
-	.map(([key, value]) => ({
-		name: value.name,
-		description: AbilitiesText[key].desc || AbilitiesText[key].shortDesc,
-		shortDescription: AbilitiesText[key].shortDesc,
-	}));
-writeFile('abilities', abilities);
-
-const items = Object.entries(Items)
-	.filter(
-		([key, value]) =>
-			!value.isNonstandard ||
-			value.isNonstandard === 'Past' ||
-			value.isNonstandard === 'Unobtainable'
-	)
-	.map(([key, value]) => ({
-		name: value.name,
-		description: ItemsText[key].desc,
-		owners: value.itemUser,
-	}));
 writeFile('items', items);
 
-const WEAKNESS = { 0: 1, 1: 2, 2: 0.5, 3: 0 }; // translate weakness ratio
-const types = Object.entries(TypeChart).map(([key, value]) => ({
-	name: key,
-	weaknesses: Object.entries(value.damageTaken).map(([key, value]) => ({
-		name: key,
-		ratio: WEAKNESS[value],
-	})),
-}));
 writeFile('types', types);
 
-const moves = Object.entries(Moves)
-	.filter(([key, value]) => !value.isNonstandard || value.isNonstandard === 'Past')
-	.map(([key, value]) => ({
-		name: value.name,
-		category: value.category,
-		description: MovesText[key].desc || MovesText[key].shortDesc,
-		shortDescription: MovesText[key].shortDesc,
-		power: value.basePower,
-		pp: value.pp,
-		accuracy: value.accuracy === true ? null : value.accuracy,
-		type: value.type,
-	}));
 writeFile('moves', moves);
 
-const createDiscriminant = ({name,baseStats,types,abilities}) => JSON.stringify({name,baseStats,types,abilities})
-
-const pokemons = Object.entries(Pokedex)
-	.filter(([key, value]) => !FormatsData[key] || pokemonIsStandard(FormatsData[key]))
-	.reduce((accumulator, [key,value]) => {
-
-		return {...accumulator, [createDiscriminant(value)]:{
-			...value,
-			gen: [LAST_GEN]
-		}}
-	},{});
-
-const mods = (gen) => {
-
-	const { FormatsData } = require(`./pokemon-showdown/.data-dist/mods/gen${gen}/formats-data`)
-	const cleanedFormatsData = Object.keys(FormatsData).reduce((accumulator,key) => {
-		
-		if(pokemonIsStandard(FormatsData[key]))
-		{
-			const FormatsDataPokemon = FormatsData[key]
-			return {...accumulator, [key] : FormatsDataPokemon}
-		}
-		return accumulator
-
-	},{})
-	
-	let ModPokedex = null;
-	if(gen != 3){
-		const { Pokedex } = require(`./pokemon-showdown/.data-dist/mods/gen${gen}/pokedex`)
-		ModPokedex = Pokedex
-	}
-
-	return { ModFormatsData: cleanedFormatsData, ModPokedex }
-}
-
-/**
- * Remove incompatible abilities' parameters
- * regarding a specific gen
- * @param {number} gen 
- * @param {{}} object 
- * @returns 
- */
-const cleanAbilities = (gen, object) => {
-
-	if(object['abilities']){
-		if(gen < 5){
-			delete object['abilities']['H']
-		}
-		if(gen < 3)
-		  delete object['abilities']
-	}
-
-	return object
-}
-
-let modsByGen = {}
-
-for(let gen=LAST_GEN; gen > 0; gen--)
-{	
-	modsByGen[gen] = {}
-	if(gen == LAST_GEN)
-	{
-		modsByGen[gen]['Pokedex'] = Pokedex
-		modsByGen[gen]['FormatsData'] = FormatsData
-	}else if(gen != 3){
-		const { ModFormatsData, ModPokedex } = mods(gen)
-		modsByGen[gen]['Pokedex'] = ModPokedex
-		modsByGen[gen]['FormatsData'] = ModFormatsData
-	}else{
-		const { ModFormatsData } = mods(gen)
-		modsByGen[gen]['FormatsData'] = ModFormatsData
-	}
-	
-}
-
-const findInheritedPokemonGenProperty = (gen,pokemonName, property) => {
-
-	for(let _gen = gen; gen < LAST_GEN; _gen++)
-	{
-		let nextPokemonGen = null;
-		if(_gen == 3)
-			nextPokemonGen = modsByGen[4]['Pokedex'][pokemonName]
-		else
-			nextPokemonGen =  modsByGen[_gen]['Pokedex'][pokemonName]
-		
-		if(nextPokemonGen)
-		{
-			if(nextPokemonGen[property]){
-				return nextPokemonGen[property];
-			}
-		}
-	}
-}
-
-for(let gen=LAST_GEN-1; gen > 0; gen--)
-{
-	Object.entries(modsByGen[gen]['FormatsData'])
-		  .filter(([key,object]) => !modsByGen[gen]['FormatsData'][key] || pokemonIsStandard(!modsByGen[gen]['FormatsData'][key]))
-		  .forEach(([key,object]) => {
-
-				if(key != 'missingno' && modsByGen[LAST_GEN]['Pokedex'][key]){ //missingno is Custom in 8th gen however, it is Unobtainable in 1st gen
-					
-					/**
-					 * We retrieve the latest pokemon's gen
-					 * And we modify it with the correct paramters according to the other gens
-					 * (no abilities, no fairy type etc.)
-					 * 
-					 * Because the pokemon object has nested objects
-					 * We must make a DEEP COPY ot it
-					 */
-					
-					const lastGenPokemon = JSON.parse(JSON.stringify(modsByGen[LAST_GEN]['Pokedex'][key]))
-					
-					// Will check and fetch values of next gen (smogon system uses reverse inheritence, example : gen1 inherit values of gen2)
-					const inheritedPokemonInfo = { 
-						baseStats: findInheritedPokemonGenProperty(gen,key,'baseStats'),
-						abilities: findInheritedPokemonGenProperty(gen,key,'abilities'),
-						types: findInheritedPokemonGenProperty(gen,key,'types')
-					}
-					
-					const richGenPokemonObject = cleanAbilities(gen,Object.assign(lastGenPokemon, inheritedPokemonInfo))
-					const discriminant = createDiscriminant(richGenPokemonObject)
-
-					if(pokemons.hasOwnProperty(discriminant))
-						pokemons[discriminant]['gen'].push(gen)
-					else {
-							
-						pokemons[discriminant] = richGenPokemonObject
-						pokemons[discriminant]['gen'] = [gen]
-					}
-
-				}
-				
-	})
-		
-}
-
-const resultPokemons = Object.values(pokemons).map((value) => ({
-	name: value.name,
-	type_1: value.types[0],
-	type_2: value.types.length > 1 ? value.types[1] : null,
-	hp: value.baseStats.hp,
-	atk: value.baseStats.atk,
-	def: value.baseStats.def,
-	spa: value.baseStats.spa,
-	spd: value.baseStats.spd,
-	spe: value.baseStats.spe,
-	ability_1: value.abilities ? value.abilities[0] : null,
-	ability_2: value.abilities && value.abilities.length > 1 ? value.abilities[1] : null,
-	ability_hidden: value.abilities && value.abilities['H'] ? value.abilities['H'] : null,
-	weight: value.weightkg,
-	baseForm: value.baseSpecies ? value.baseSpecies : null,
-	prevo: value.prevo ? value.prevo : null,
-	gen: value.gen.sort()
-}))
-
-writeFile('pokemons', resultPokemons);
-
-const learns = [];
-Object.entries(Learnsets)
-	.filter(([key, value]) => !FormatsData[key] || pokemonIsStandard(FormatsData[key]))
-	.forEach(([key, value]) => {
-		if (value.learnset) {
-			Object.keys(value.learnset).forEach(move => {
-				learns.push({
-					pokemon: PokedexText[key] ? PokedexText[key].name : key,
-					move: MovesText[move] ? MovesText[move].name : move,
-				});
-			});
-		}
-	});
 writeFile('learns', learns);
 
-const natures = Object.values(Natures).map(value => {
-	const nature = { name: value.name };
-	if (value.plus) nature[value.plus] = 1;
-	if (value.minus) nature[value.minus] = -1;
-	return nature;
-});
 writeFile('natures', natures);
 
-const pokemonTier = Object.entries(FormatsData)
-	.filter(([key, value]) => pokemonIsStandard(value))
-	.map(([key, value]) => ({
-		pokemon: PokedexText[key] ? PokedexText[key].name : key,
-		tier: value.tier ? removeParenthesis(value.tier) : undefined,
-		technically: value.tier ? value.tier.startsWith('(') : false,
-		doublesTier: value.doublesTier ? removeParenthesis(value.doublesTier) : undefined,
-	}));
 writeFile('pokemonTier', pokemonTier);
